+++ src/VaultSharp/V1/SecretsEngines/OpenLDAP/OpenLDAPSecretsEngineProvider.cs
@@ -1,0 +1,25 @@
+﻿using System.Net.Http;
+using System.Threading.Tasks;
+using VaultSharp.Core;
+using VaultSharp.V1.Commons;
+
+namespace VaultSharp.V1.SecretsEngines.OpenLDAP
+{
+    internal class OpenLDAPSecretsEngineProvider : IOpenLDAPSecretsEngine
+    {
+        private readonly Polymath _polymath;
+
+        public OpenLDAPSecretsEngineProvider(Polymath polymath)
+        {
+            _polymath = polymath;
+        }
+
+        public async Task<Secret<StaticCredentials>> GetStaticCredentialsAsync(string roleName, string mountPoint = SecretsEngineDefaultPaths.OpenLDAP, string wrapTimeToLive = null)
+        {
+            Checker.NotNull(mountPoint, "mountPoint");
+            Checker.NotNull(roleName, "roleName");
+
+            return await _polymath.MakeVaultApiRequest<Secret<StaticCredentials>>("v1/" + mountPoint.Trim('/') + "/static-cred/" + roleName.Trim('/'), HttpMethod.Get, wrapTimeToLive: wrapTimeToLive).ConfigureAwait(_polymath.VaultClientSettings.ContinueAsyncTasksOnCapturedContext);
+        }
+    }
+}